Add exception handling for build failure
Prior to this commit, we did not have exception handling in place
for build failures.  This meant that the workspace was not getting
properly cleaned up on a failed build, so the build machines' storage
was getting clogged with workspaces from failed builds.  This
commit adds exception handling so that if the build fails in the
checkout or build stage, the workspace will be deleted.