Bugfix/add common h (#38)
* Add common.h file and update dependancies
Prior to this commit various macros were duplicated in the steering, throttle, break and obd_can firmware. This commit adds a high level common.h file and moves these duplicate macros to this common location. It also renames can_gateway.io which was misspelled.

* Update common.h dependancies for tests
Prior to this the test files in this project used their own local macros instead of the ones defined in common.h. This commit removes those local macros and replaces them with an include from common.h.

* Add common.h file
Prior to this commit, the common.h file which holds common macros in multiple firmware files, did not exist. This commit adds this common.h file.

* Add common.h file
Prior to this commit, the common.h file which holds common macros in multiple firmware files, did not exist. This commit adds this common.h file.