Standardize the PID library (#42)
* Remove PIDV1c library from throttle control
Prior to this commit the PIDV1 library was unneccessairly included in the throttle control code. It is not used so is now removed.

* Move new PID library, remove old PID library
Prior to this commit the custom PID library, used in the steering controller, was located in the steering controller directory. This commit moves it to the arduino_libraries directory so that it can be referenced by multiple firmware projects. In addition, the PIDv1 library was removed so that it could be replaced with the new custom library.

* Add safety check for pid_update calculation
Prior to this commit there was no checking to make sure that the delta t being passed into the PID update calculation was a valid value. This commit adds functionality to make sure this value is valid before running the PID update.

* Add new PID library, cleanup, minor safety changes
Prior to this commit the old PIDv1 library was implemented to do control. This commit removes the code to interact with the old library and adds new code to interact with new custom PID library. In addition there was minor cleanuip done related to the PID controller, and a couple minor safety bug fixes were made, primarily to error check PID calculation and to ensure that the system properly shifts out of enabled mode when there is a timeout in receiving control command messages.

* Remove serial debugging, depreciated PID averaging
Prior to this commit the processSerialByte function received commands over serial that were used for debugging. The purpose of this code is unclear to the user and it looks sloppy. This commit removes this code. In addition this commit removes code that was used to do averaging of the steering PID controller ouput but is now depreciated.