Add smoother control of steering, brake (#55)
* Add smoother control of steering, brake
Prior to this commit steering, throttle and brake control didn't have any filtering of their respective setpoints. Further, steering lacked the functionality to smooth control via dynamically changing the max steering rate. This commit adds the functionality to smooth steering and brake through exponential filtering. It also adds the functionality to smooth throttle control through filtering, but does not implement this (as it was not neccessary). Finally it adds dynamic control of the max steering wheel rate, which further smooths out control.

* Add style changes
Prior to this commit there were a couple if statements that were not enclosed in curlies, thus didn't follow the coding standard. This commit adds those curlies.