Changed the loop timer expiration and detection of driver input
Changed how the loop timer decides when to execute from using
micro to milliseconds and added filtering code for the torque inputs.
The check driver steering override function returns a boolean
value that indicates if the driver is holding the steering wheel
or at least that the control thinks the driver is.