Refactor to conform to coding standards
Prior to this commit, the code in the brake module did not conform to
the coding standards in contributing.md.  This commit corrects some of
those issues to bring it closer to conforming to the standard.  These
edits are not complete and there is still further work to be done.