Add verbose description for max steering velocity
Prior to this commit, the STEERING_COMMAND_MAX_VELOCITY_FACTOR did not include a verbose description string. Changing this factor is the recommended way to actuate smoother steering control at the lower level. Thus this description string needed to be updated to make this more clear for users. This commit adds a more verbose description for this factor.