Update the README to reflect V1 hardware status
Prior to this commit, the README did not have up-to-date references
to the board design status. This commit updates the "Boards" section
to point to the directory in OSCC repo with the board schematics and
design file resources that are required to to fabricate the V1.0.0
sensor interface and actuator control boards.