Remove steering and throttle ADC diagnostics (#79)
Prior to this commit ADC diagnostics were included as part of the throttle and steering modules. After extensive testing with these diagnostics tools it was determined that more in-depth system level testing was required in order to properly determine threshold values which trigger an ADC diagnostic override. Rather than burden the devel branch with diagnostics tools which may sometimes trigger false positives we are electing to remove this functionality until further testing is done. This commit removes the code that does ADC diagnostics in the throttle and steering modules.