Feature/add report msg verbosity (#99)
* Add verbosity to brake report message
Prior to this commit, the brake report message was not using all of the fields available to provide info to the user. This commit adds the functionality to provide more information about what is going on in the brake module by using more of the fields in the report message.

* Add verbosity to steering report message
Prior to this commit, the steering report message was not using all of the fields available to provide info to the user. This commit adds the functionality to provide more information about what is going on in the steering module by using more of the fields in the report message.

* Add verbosity to throttle report message
Prior to this commit, the throttle report message was not using all of the fields available to provide info to the user. This commit adds the functionality to provide more information about what is going on in the throttle module by using more of the fields in the report message.