Add logic to detect OSCC module disable
Added logic into the oscc_interface.c and commander.c files to detect
whether or not the driver is trying to take control of the vehicle.
Reworked the logic in the commander.c to set a precendent for disabling
the OSCC modules before checking if the enable button is set or trying
to command the modules to a specific set point.