Add commander high and low frequency functions
Changed the calls to the commander to consist of both high and low
frequency functions.  This was specifically to separate CAN reads
from the slower update functions.  Also removed the commander_data
structure as it turns out the data in it could be declared statically
local to where it is being used.  This forced a change in how to
identify when the commander is enabled or disabled.