Add socketCAN support into joystick commander
Multiple additions to support socketCAN: Changed the build system to use
cmake instead of the Makefile.  The cmake can take a command-line parameter
to select linuxCAN or socketCAN on the build (e.g. cmake OSCC_INTF=linuxcan ..)
Changed the name of the oscc_interface.c to oscc_linuxcan.c and added an
oscc_socketcan.c file to allow for support for either of the interfaces.