Update include file, fix typo, add Build instructions
The commander.c file used canlib.h from the Kvaser installation.  This had to
be changed to stdint.h so that commander is independent of the Kvaser install.
The oscc_socketcan.c file printed out the name of the can channel it is
opening but did not have a carriage-return at the end of the line.
Added build instruction on how to use the command-line cmake to build
the socketcan and linuxcan versions of the joystick-commander.