Update to reflect contributing standards
Prior to this commit, some of the files in the CAN gateway folder did
not conform to our contributing standards.  This was mostly formatting
of parentheses, however there were a couple of if/else if statements
that were missing the terminating else statement per 12.7.  This commit
adds in the terminating else statements, but with a TODO comment to determine
the default behavior of these statements if the other conditions are not met.