Update to conform to contributing guidelines
Prior to this commit, there were some files in the steering controler
that did not conform to our contributing guidelines, mostly stylistic
things like spaces between parentheses.  There is still some commented-out
code that needs to be discussed before it is deleted, since our contributing
guidelines state that there is to be no commented-out code.