Refactor for testing
Prior to this commit, the functions were all in the .ino file without
prototypes in a header file.  This caused difficulties for keeping
tests separate from code.  This commit is the first part of the refactoring
process that is pulling out the necessary functions into a .h and .cpp to
enable future testing and maintenance.