Add warning to steering control state struct
Prior to this commit, there was no warning for the PID numbers in the
current control state struct for the steering control module.  Users
have been faulting their cars by possibly making changes to these
numbers.  This commit adds in the warning in an attempt to lower the
number of faults by including a warning.