Move PID assignment from loop into setup
Prior to this commit, the PID assignment was taking place in the loop function.
These values will never change, and by having the assignment in the loop
function, it was being done every time the loop was run, when it only needed
to be done once.  This commit moves those assignments into the setup function
so that they are done once instead of multiple times needlessly.