Add test and warning for altered PID values
Prior to this commit, there was no warning when users changed the
steering PID values. Changing the PID values could cause the
vehicle to fault.  This commit adds a test into the setup function
and prints a warning to the screen that the user has changed the
PID values and this may cause a fault.