Add exit if test fails
Prior to this commit, there was no exit in loop if the test failed,
so the user was not required to pass the test to run their software.
This commit puts an exit on failed test into the code so that users
have to return the PID values back to the original values before
they are able to run their code.