Rename "control" dir to "utils"
Prior to this commit, the "control" directory contained
joystick_commander (which is for controlling) and diagnostics_tool
(which is not for controlling), so the directory name "control" was not
representative of the contents. This commit renames the directory to
"utils" to reflect that it contains utilities for interacting with
OSCC, but is not firmware.