Remove override status if control disabled
Prior to this commit, the operator override status flag would be set
upon override but would then never be unset because the override logic
was only carried out when control was enabled, but after override
control is disabled. This commit moves the operator override disabling
logic outside of the control enabled check.