Do not enable if operator override in effect
Prior to this commit, it was possible for an operator override to occur,
and then for control to be immediately re-enabled afterward if an old
message with an enable command was received (before the controller had
received the override report and sent a disable command). This commit
ensures that control can only be enabled if the operator override flag
is not set, and it will check for an operator override if either control
is enabled or the override flag is set (so that it can eventually clear
it).