Check for OBD timeout and disable if one occurs
Prior to this commit, the steering and brake modules did not check for
an OBD timeout which allowed them to operate as if they were receiving
OBD information (required for their feedback systems to work correctly)
when they in fact were making control decisions on invalid information.
This commit adds checks for the validity of the data that the modules
receive from the Chassis State 1 report (sent from the CAN gateway)
before using the data, and keeps a timeout tracker of when they receive
valid data. If valid data is not received after a certain amount of
time, they disable and set a new OBD timeout fault in their report messages,
which is checked by the controller. Once valid OBD data begins to be
received again, the fault is cleared and the controller can re-enable
control.