Use floats in steering and throttle control fields
Prior to this commit, the fields of the module's control state
structure were using data types that matches what would eventually be
sent out in the report, but this lost some precision that made the PID
controller operate more smoothly. This commit uses floats for those
fields and casts them to the appropriate data types during report
publishing.