Move old brake test to brake utils
Prior to this commit, there was brake_controller_r0_test in the brake
tests directory, but this was not actually a test but instead a utility
that parses messages sent over on the serial interface and actuates the
brake system (opening and closing solenoids, turning things on and off)
which is used to verify and diagnose brake functionality. This commit
moves the utility to the utils directory and renames it to
"serial_actuator" to reflect its purpose.