Update brake PID values
Prior to this commit, the PID values used in the brake controller did
not allow for the brakes to fully disengage (lowest pressure was not low
enough) and did not brake hard enough (highest pressure was not high
enough). This commit changes the PID values to allow for harder braking
and fully disengaged brakes.