Rename write_sample_averages_to_dac
Prior to this commit, the function was not named in a way that explained
what was going on which required a comment to be included with the
function when it was called in the steering and throttle modules. This
commit renames the function and adds a better description to its header
comment so that the function call can stand on its own.