Use signal smoothing library in steering module
Prior to this commit, the steering module was using an exponential
moving average on the values read from its torque sensor, but it was
taking place in the operator override section instead of where the
actual torque sensor reading was taking place. This commit moves the
filtering into the read_torque_sensor function, and also uses the new
signal smoothing library function.