Use signal smoothing library in brake module
Prior to this commit, the brake module was using an exponential
moving average on the values read from its master cylinder pressure
sensor, but it was taking place in the operator override section.
This commit creates a function for reading the master cylinder pressure
and moves the filtering into it. It also uses the new signal smoothing
library function.