Change raw_adc_to_pressure input to int
Prior to this commit, the input of raw_adc_to_presure was a uint16_t
which doesn't make sense because the function is fed the output of
analogRead which returns an int. Furthermore, in the function the value
is immediately cast to a float, so there was no need for it to ever be
cast to a uint16_t. This commit changes it to expect an int.