Update brake control PID values
Prior to this commit, the brake control behavior did not match its
previous behavior due to the addition of the exponential filters on the
sensor reads changing the response of the system. This commit modifies
the PID terms to make the PID loop behave in a way similar to before the
exponential filters were added.