Restore make monitor command
Prior to this commit, the `make monitor` command from the previous
Makefile-based build system was missing which led to painful monitoring
of serial output by manually invoking `screen` every time. This commit
adds the custom target `monitor` which runs normal `screen`, and the
custom target `monitor-log` which runs `screen` but prints the serial
output to a file called `screenlog.0`.