Check steering sensors for valid values
Prior to this commit, the sensors used in the steering module (i.e.,
torque sensor) were not checked for valid values so it was
possible for a sensor wire to become disconnected while control
continued to be enabled. This commit does a periodic check of the
sensor values to see if any of them return a value of 0 indicating a
problem because the sensor pins are tied to ground and will be 0 if
they're disconnected.