Optimize SSD1325 library for speed
Prior to this commit, the SSD1325 library set the CS lines for each byte
rather than a block of bytes which caused slowdown in the overall speed
of the main while loop. This commit modifies the SSD1325 library to
assert its CS pin before a block of successive SPI transfers and then
deassert afterward.