Add timer for display updating
Prior to this commit, the display was updated once every time through
the loop which exceeded the processing power of the Arduino and caused
CAN frames to be received from OBD at such a slow pace that they were
considered to have timed out. This commit adds a timer to the display
update logic to only update every 250 milliseconds.