Reduce responsibilties of the CAN gateway
Prior to this commit, the CAN gateway would receive messages on the OBD
bus and then repackage them into "chassis state" messages, and also
publish heartbeat messages with chassis state information. This commit
removes that overhead by instead having the CAN gateway look for
specific OBD messages on the OBD bus and then republish any that it
receives to the Control CAN bus for further consumption by the modules
without any processing or bundling up into "chassis state" reports.

This allows the gateway to simplify its role by not doing any processing
of OBD messages but instead simply acting as a bridge between the OBD
bus and the control bus, and removes the need for it to keep track of
timeouts and health information as that can be done by the modules
themselves when they look for the specific OBD messages that they care
about.