Use timer for steering fault checking
Prior to this commit, the steering module used timestamps to keep track
of when to check for faults and when timeouts occurred. This commit
replaces the timestamp system with a timer system where faults are
checked ten times a second. For command timeouts, a flag is cleared
every time a command is received and set at the end of the interrupt
service routine - if a command is not received between one ISR call and
the next (a span of 100ms), then the timeout flag has not been cleared
and a timeout is considered to have occurred. For sensor validity
checks, they are run once with every ISR call, meaning they're run
every 100ms.