Restore PID controller to brake firmware
Prior to this commit, the PID controller had been removed from the brake
firmware in expectation that the brakes would be controlled from a
higher level application. However, due to the way the brake actuator
operates, it isn't feasible to have the brake firmware be stupid and
controllable from higher up; instead, to maintain a consistent API in
future versions where brake actuation is done differently, the PID
controller is restored to the firmware.