Add magic number to reports and commands
Prior to this commit, it was possible for OSCC commands and reports to
conflict with OBD messages if an OBD message was forwarded from the OBD
CAN bus to the Control CAN bus and that message had an ID matching an
OSCC message. This commit adds a magic number field to commands and
reports that must be set when sending onto the Control CAN bus and
checked when processing. This prevents the situation where a message is
received with an ID matching an OSCC message and then processed as if it
were that message resulting in undesired behavior.