Use two bytes for magic sequence
Prior to this commit, the magic number used to distinguish OSCC CAN
frames was a single uint16_t. This commit changes the magic number to be
two uint8_ts (0x05, 0xCC) so that the API and the firmware can both
check those two bytes first before casting the payload to a report type.
On the API side, this allows it to distinguish immediately upon
receiving a frame whether it's from OSCC (if so, the report callback can
be called), or an OBD frame (if so, the OBD callback can be called). On
the firmware side, it allows the firmware to quickly ignore OBD frames;
there are many of them and the faster they can be ignored the better.