Reduce fault checking frequency
With the addition of OBD messages on the control CAN bus, the time per
main loop iteration increased due to the time required to check the ID
of each OBD frame on the bus. This commit reduces the frequencies for
each module (brake is even further reduced due to its additional
computational requirements due to its PID loop).