Move OSCC magic bytes to a separate header
Prior to this commit, the OSCC magic bytes were in oscc.h which was
being included by both the application layer and the firmware layer,
which caused problems because the firmware missing headers that oscc.h
needed to include. This commit moves the magic bytes into a separate
header that is included only by each CAN protocol header, removing the
need for magic.h to ever be included directly.