Remove Kia Soul references from the firmware
Prior to this commit, the firmware had a concept of what vehicle it was
building for which is no longer in line with the way the software stack
is architected (i.e., firmware should be consistent across vehicles, all
that changes is inside of the vehicle specific header like
"kia_soul.h"). This commit removes the KIA Soul references and keeps the
firmware vehicle agnostic.