Remove joystick commander
Prior to this commit, joystick commander existed in the main OSCC repo
which caused versioning problems where joystick commander might need to
be updated but a version bump in the firmware/API was unwanted. This
commit removes joystick commander from the main repo, but it lives on in
its own separate repo.