Add patch for legacy actuator board support
Prior to this commit, the new firmware would not work on older (< 1.1.0) versions of the actuator control board due to new startup tests that use new pins. This commit adds a CMake option to allow legacy board users to disable these startup checks, allowing them to still use the firmware as usual. It also updates the README to reflect these changes.