add instructions for signing individual commits
Prior to this commit the contributing guidelines only illustrated how to
sign merge commits, but not the individual commits. This commit adds
instructions with, tactically, how to configure git to sign all commits;
however, no instructions on how to generate the GPG keys themselves.
This commit does include a link to GitHub for further instructions for
configuring git and generating GPG keys.