Rust extern globals don't usually need #[link_name].
In Rust, `static` items in an `extern` block are not subject to name
mangling. So the name given to them at link-time is the same as the name
in the source code, by default. The `#[link_name]` attribute is only
necessary if the link name should be different than the name in the
source code, which it was not in any of these cases.