Update PBT for new steering
Prior to this commit, PBT weren't distinguishing between spoof signal low range and spoof signal high range, since this had only been added in as part of this branch. This commit converts the PBT to use the new low and high min/max values. It also updates the PBT to check the analogRead value shl 2 places, since we have changed that in the firmware as well.