Remove shifting from analogRead
Prior to this commit, I had re-added shifting to the analog read function in the steering firmware, so that the read-in torque values more closely resemble our output spoofs and correspond better to volts (since torque sensor / STEPS_PER_VOLT will provide the analogRead value in volts if the value is shifted first). This commit removes it for now, since it would make more sense to do this on a different branch to all of the firmwares if we decide to do that.