Update unit tests to work with low/high min/max
Prior to this commit, the unit tests were still using the old spoof value range, which did not differentiate between high and low signal limits. This commit adds that functionality into the unit tests so it can be consistent with the new spoof min/max values on both low and high signal ranges.