Use enums for module DTCs
Prior to this commit the DTCs were macros, but the gateway display needs
to know how many DTCs each module has which requires another macro that
needs to be updated every time a new DTC is added. This commit changes
the macro to an enum instead so that the last item in the enum is always
the number of DTCs in that enum.