Remove gateway status and DTC from display
Prior to this commit, the display had gateway status on the status
screen and displayed gateway DTCs on the DTC screen, which is no longer
applicable following the recent code reorganization and rewrite which
reduces the responsibilities of the gateway to simple OBD message
forwarding. This commit removes all gateway display items.