Check for valid CAN socket before using
Prior to this commit, functions assumed that the variable can_socket was
valid and had been initialized by oscc_open, but this was not a fair
assumption. This commit initalizes can_socket to -1 and then all
functions that use can_socket check to ensure it is greater than 0
(guaranteed to be true if a successful call to socket() occurs during
oscc_init_can()) before using the socket.