Check for valid CAN interface during oscc_init_can
Prior to this commit, oscc_init_can could successfully open a CAN socket
even if the CAN interface had not been set up using the 'ip link'
commands, resulting in the erroneous belief that oscc_open was
successful. This commit attempts to write a single empty byte after
opening the socket to check for the existence of a valid interface.