Update operator override
Prior to this commit, the previous methodology for detecting operator override was no longer sufficient to allow for using the new expanded range of possible spoofed outputs.
This commit updates the previous method from reading a specific value from the torque sensor to instead detect large amounts of torque over a period of time.