Update unit tests
Prior to this commit, we only allowed tests to perform analogReads of one value. With the new way of detecting steering overrides, tests must be able to write differential signals to the mock Arduino test classes. This commit updates the Arduino mock to be an array of analogReads, allowing the steering tests to correctly detect operator overrides.