Add enable and disable messages to all modules
Prior to this commit, the ability to enable and disable a module was
part of the command messages themselves. This made for an awkward
situation where the "enable" field would be set or unset and the rest of
the fields (the command values themselves) would be invalid. This commit
gives each module their own Enable and Disable messages which, when
received by a module, would then Enable or Disable.