Update mins for throttle
Prior to this commit, the minimum low and high signal values were set at a non-zero number. Although the previous minimum values came from directly measuring voltages in their steady state, they would cause the vehicle to rev a bit as OSCC was enabled. This commit changes the minimum values to be 0, which will remove this undesired behavior.