Move static functions to internal header file
Prior to this commit, there were a lot of static functions in the API
which makes for both difficult testing (testing libraries have no way of
getting access to those functions) and clutter in the main oscc.c file.
This commit moves the static functions to a separate header meant for
internal use (src/internal/oscc.h) and removes the static keyword so
they can be seen by test libraries in the future.