Update Jenkinsfile, roll back rust target
Prior to this commit, bindgen was attempting to use the latest stable version of Rust, which was causing some build issues since it had support for things that were unsupp
orted in the build machine version, 1.16. This commit tries to build against an older rust target.
This commit also updates the Jenkinsfile to initialize the vcan interface necessary for running the API tests and includes the script necessary.