Initialize sigaction struct to zero
Prior to this commit, the sigaction struct used to register the SIGIO
signal handler on CAN frame reception was not initialized to zero which
caused occasional strange behavior due to the undefined contents of the
sa_flags and sa_mask fields of the struct. This commit initializes the
struct to zeros before use to ensure it is in a known state.