Use regex in build
Prior to this commit, we were explicitly naming each function or type we wanted to whitelist and generate bindings for. This commit changes it to instead use regular expressions, cleaning up the build code a bit. It also turns off clang mangling so that we can automatically generate bindings to the c functions in the API.