Add explicit error checking
Prior to this commit, the tests were unwrapping the results of CAN operations without checking if they succeeded or not. This would sometimes result in the tests failing due to panics if they attempted to read from the CAN bus at the same time as another test. This commit removes the calls to unwrap() and instead explicitly tests the results for success or failure, handling both cases differently.