Add a cmake configuration file
Prior to this commit the user of the OSCC api was responsible for including
defines of each available car in the oscc api. This commit removes that pain by
adding a .cmake file to the OSCC api so that the user can include it in their
own CMakeLists.txt file keeping the details of which cars the api supports
within the api.