Turn off steering override
Prior to this commit, steering override was enabled by default. Since it is difficult to override the steering module by grabbing a moving steering wheel, and instead is much safer to use the E-stop or brake disabling methods to terminate OSCC control, the steering override system must instead be opted into. This commit adds a CMake option that enables users to include steering override detection if they choose to, and displays a warning as to why it is no longer enabled by default.