Update steering override threshold
Prior to this commit, the steering override threshold was at 1600, which could potentially cause an override to occur when either the clockwise or counter-clockwise mechanical limits of the steering rack were hit, or if control quickly switched back and forth. This commit raises that value to 1900, which requires slightly more operator effort but produces less false positive override scenarios.