Move toolchain override to cargo command
Prior to this commit, we were attempting to set the toolchain override from the Jenkinsfile, which wasn't always happening. This commit moves the override directive to the `cargo test` command that is run within CMakeLists.txt, ensuring that the toolchain is overridden and rustc v 1.17.0 is used to run the tests.