Add seperate dirs for each vehicle's API tests
Prior to this commit, each stage of vehicle testing would attempt to create and enter a `build` directory in the API folder. Since the cleanup doesn't occur until after all stages are completed, this commit changes the build dirs to be vehicle specific to avoid them trying to create and run property tests from the same directory.