Remove read timeout on test CAN socket
Prior to this commit, there was a 1 second read timeout set for the vcan sockets being opened by the tests. This would cause issues sometimes where a socket had gone out of scope but was still owned by a finished test, causing the next test to be unable to open its own socket. This commit removes that since it isn't necessary.