Code cleanup, changing rust versions working
Prior to this commit, there were a bunch of commits as I tried to get the rustup version to change to v1.17.0 so that bindgen could compile, and therefore we could run the tests on the build machine. This commit cleans up all of the junk and sets up one nice commit where we are at a working state with that part of the test.