Run all Rust commands in 1.17.0 shell
Prior to this commit, the Jenkinsfile was attempting to run the rust commands in the same shell. This was causing weird issues for rust's environment variables, even when the rust version override was being set. This commit runs the rust testing commands in a shell using the required rust environment.