Use rustup shorthand toolchain override
Prior to this commit, there was some amount of black magic necessary in order to change the rustc toolchain version for running the api property based tests. This commit adds an environment variable to the cargo directory, allowing for use of the shorthand toolchain specifier and doing some code cleanup in the CMake and Jenkins files. It also un-ignores a currently failing steering override test, since we will need to address and fix any failing test cases as they occur.