Constrain from brake tests, update min high/low for EV
Prior to this commit, we were not constraining input to our model function of calculating the brake spoof values for the EV. This resulted in weird behavior for negative inputs, even though the OSCC API was constraining bw 0 and 1.0 appropriately. This commit updates the pbt for the ev brake tests to constrain inputs similarly, allowing us to get valid output. It also updates the min high/low signals for the brake spoofs to be the floor of the brake spoof calculation since C will truncate when casting a float to an int.