Remove check for controller command timeouts
Prior to this commit, there were checks to ensure that commands were
received on a consistent basis when enabled, which often led to
erroneous timeouts due to timing limitations in the Arduino hardware and
CAN communication. The timeout check is unnecessary because the modules
won't do anything if no commands are given to them. This commit removes
the controller command timeouts.