Move fault message into a global messages file
Prior to this commit, the fault message got its own file with the single
fault report message inside of it. This commit creates a global CAN
protocol file that contains the fault report for now but will contain
other global messages in the future. This keeps them all within the same
range, allowing for easy filtering.