Add CAN filtering for module-specific messages
Prior to this commit, there was no filtering being done on received CAN
messages by the CAN transceiver, so the MCU had to service every CAN
frame that appeared on the bus. This commit adds filtering to each
module so that it only sees messages that are pertinent to its
operation, reducing MCU load.