Add software vehicle CAN and improve io signals
Prior to this commit the vehicle CAN was forwarded through hardware only. Also
used the older signal handler rather than the sigaction callback leaving the
signals prone to a known race condition in the signal handler. This commit
adds the initial layout and framework for supporting vehicle CAN through a
device on the computer rather than forwarding from the OSCC hardware. This
commit also moves to the sigaction callback rather than the signal handler and
uses the thread id incase the API is called within a thread for improved
reliability.