Incorporate a template for raising issues.
Prior to this commit, when raising a new issue users are presented with a blank canvas.

This leads to inconsitent reports from users as well the requirement for the maintainers
to poll for information that we generally always need up front. As a solution,
Github facilitates templating the default issue dialogue.

This PR represents
an attempt at templating the issue dialogue so that users are prompted to provide
information that would be helpful to the maintainers from the outset.

For an example of another project that employs a helpful issue template,
see [rust-bindgen](https://github.com/rust-lang-nursery/rust-bindgen/issues).
Click the green "New Issue" button to see a great template in action.