Add getting started resources to API.
Prior to this commit there was no real indication of how to incorporate the API into your own project or a step by step for a complete example without digging into an external repository like the joystick commander.

This PR represents an attempt to make the OSCC API self-explanatory so that a new user is able to get up and running without needing to go anywhere else.

Becasuse I'm new to the REPO and the functionality I expect change requests on this PR to help fill in the docs. I'd appreciate reviewers providing specific sentences for me to work with in those requests.

After this commit, the intention is that a new user can visit the OSCC `api` directory and have the resources they need to start their own project that incorporates the API.

Thanks for taking a look!