Move fault detection to operator override
Prior to this commit the firmware would cause a pause in the reporting messages
due to interrupts. It was determined that moving the fault detection to where
the operator override detection occurs would remove the pause due to a blocking
interrupt and reduce the number DAC reads required by the firmware. It was also
determined that leaving the reporting message as a timer interrupt yielded the
most desirable results for maintaining the desired reporting frequency.