Fix git merge
Prior to this commit the git merge moved all of the global_can_protocol.h back
to fault_can_protocol.h which was not the desired state. This commit fixes that
by putting everything back to the global_can_protocol.h and making a wrapper
header so that there are no breaking API changes.