Add version to report messages
Prior to this commit the can_protocols did not have a place to produce CAN
messages. This commit changes that by adding the version and a bit for if the
git repo has a clean or dirty tag. However the rust property tests stop working
because bindgen bitfields have open issues.